+++ Frontend/e_shop/src/components/context/user_context.js
@@ -1,4 +1,5 @@
 import { useState,useEffect,createContext} from "react";
+import { mockGetUsers,mockChangeStatus,mockDeleteUser,mockAdminChangePassword} from "../Mockdata_services/UserService/UserService";
 import axios from "axios";
 export const users_context = createContext();
 
@@ -6,14 +7,24 @@ export function UsersProvider({children}){
 
     const [users,setUsers] = useState([]);
     const users_api = process.env.NEXT_PUBLIC_DB_API_USERS;
+    const mock_service = process.env.NEXT_PUBLIC_MOCK_SERVICE;
+    
  
   
 
 
     async function getusers(){
         try{
-            const users = await axios.get(`${users_api}/getallusers`,{withCredentials:true})
-            setUsers(users.data)
+            if(mock_service === 'true'){
+                const users = await mockGetUsers()
+                setUsers(users)
+
+            }
+            else{
+
+                const users = await axios.get(`${users_api}/getallusers`,{withCredentials:true})
+                setUsers(users.data)
+            }
             
         }
         catch(error){
@@ -24,7 +35,13 @@ export function UsersProvider({children}){
 
     async function changeStatus(data){
         try {
-            await axios.patch(`${users_api}/changeStatus`,data,{withCredentials:true})
+            if(mock_service === 'true'){
+                await mockChangeStatus()
+            }
+            else{
+                await axios.patch(`${users_api}/changeStatus`,data,{withCredentials:true})
+
+            }
             await getusers()
         } catch (error) {
             return error;
@@ -34,7 +51,13 @@ export function UsersProvider({children}){
 
     async function deleteuser(id){
         try {
-            const res = await axios.delete(`${users_api}/deleteUser/${id}`)
+            if(mock_service === 'true'){
+                await mockDeleteUser(id)
+            }
+            else{
+                const res = await axios.delete(`${users_api}/deleteUser/${id}`)
+
+            }
             await getusers();
         } catch (error) {
             return error;
@@ -45,7 +68,13 @@ export function UsersProvider({children}){
 
     async function admin_change_password(data){
         try {
-            const res = await axios.patch(`${users_api}/updatePassbyAdmin`,data,{withCredentials:true})
+            if(mock_service === 'true'){
+                const res = await mockAdminChangePassword(data)
+            }
+            else{
+                const res = await axios.patch(`${users_api}/updatePassbyAdmin`,data,{withCredentials:true})
+
+            }
             
         } catch (error) {
             return error;
